+++ lib/scrobblenaut.dart
@@ -14,3 +14,4 @@ export 'package:scrobblenaut/src/scrobblenaut.dart';
 // Extensions.
 export 'package:scrobblenaut/src/extensions/album_extension.dart';
 export 'package:scrobblenaut/src/extensions/artist_extension.dart';
+export 'package:scrobblenaut/src/extensions/tag_extension.dart';